Revert "api_server: add token verification logic"
This reverts commit 81f3de7523de77df62497c2a787fd0e6e2c21c78.
